requirements.txt
flask
flask_sqlalchemy
google-generativeai
python-dotenv
